シンボル処理別に作成
- コメントを # に変更。
- 変数制限に : を追加。

+++ Classes/Interpreter/QBasicFunctionEntity.cpp
@@ -39,8 +39,8 @@ QBasicFunctionEntity::QBasicFunctionEntity(
 										   const vector<QBasicVariableEntity> &argNames,
 										   const VariableType returnType,
 										   const vector<VariableType> returnSubTypes,
-										   const long startOffset,
-										   const long endOffset) {
+										   const int startOffset,
+										   const int endOffset) {
 	this->name = name;
 	this->argNames = argNames;
 	this->returnType = returnType;
