+++ Classes/Interpreter/QBasicFunctions.h
@@ -38,8 +38,8 @@ public:
 									const vector<QBasicVariableEntity> &variableTypes,
 									const VariableType returnType,
 									const vector<VariableType> returnSubTypes,
-									const long startOffset,
-									const long endOffset);
+									const int startOffset,
+									const int endOffset);
 	
 	/**
 	 * 名前で存在判定
